pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = "Agentic RAG product: scraper, ingestion pipeline, and chatbot on Ollama"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "truststore>=0.10.4",
]

[dependency-groups]
dev = [
    "ruff>=0.15.0",
]

[tool.ruff]
target-version = "py313"

[tool.ruff.lint]
select = ["F", "B", "UP", "I"]
